requirements.txt
# Agent framework
crewai
crewai-tools
litellm

# UI
streamlit

# Environment variables
python-dotenv
